Updated posts apis w.r.t user association

+++ src/posts/posts.service.ts
@@ -1,6 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { PostDto } from './dto/post.dto';
+import { PostDto, CreatePostDto } from './dto/post.dto';
 
 @Injectable()
 export class PostsService {
@@ -10,19 +10,41 @@ export class PostsService {
     return this.prisma.post.findMany();
   }
 
-  async findOne(id: number) {
-    return this.prisma.post.findUnique({ where: { id } });
+  async findOne(id: number, authorId: number) {
+    const post = await this.prisma.post.findUnique({ where: { id, authorId } });
+
+    if (!post) {
+      throw new HttpException(
+        {
+          success: false,
+          error: 'Post not found',
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    return post;
   }
 
-  async create(data: PostDto) {
-    return this.prisma.post.create({ data });
+  async create(data: CreatePostDto) {
+    const post = await this.prisma.post.create({ data });
+    return {
+      message: 'Post created Successfully',
+      data: post,
+    };
   }
 
-  async update(id: number, data: PostDto) {
-    return this.prisma.post.update({ where: { id }, data });
+  async update(id: number, data: PostDto, authorId: number) {
+    await this.prisma.post.update({ where: { id, authorId }, data });
+    return {
+      message: 'Post updated Successfully',
+    };
   }
 
-  async remove(id: number) {
-    await this.prisma.post.delete({ where: { id } });
+  async remove(id: number, authorId: number) {
+    await this.prisma.post.delete({ where: { id, authorId } });
+    return {
+      message: 'Post deleted Successfully',
+    };
   }
 }
